% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration1.R
\name{Calibration1}
\alias{Calibration1}
\title{Calibration method 1 (Normal Approximation)}
\usage{
Calibration1(x1_mean, x2_mean, x1_var, x2_var, cov, delta, alpha_p = 0.05, w = \eqn{\frac{cov}{x2_var}}, \eqn{W = x1_mean - w*ifelse(abs(x2_mean) < delta, x2_mean, 0)})
}
\arguments{
\item{x1_mean}{the sample mean of the treatment effect estimation (mean difference between the treatment arm and control arm)}

\item{x2_mean}{the sample mean that evaluate the heterogeneity between the control arm and external control arm}

\item{x1_var}{the estimated variance of x1}

\item{x2_var}{the estimated variance of x2}

\item{cov}{the covariance between x1 and x2, which is the variance of the endpoints of the control arm}

\item{delta}{the equivalence margin, which is the maximum difference that can be treated as clinically acceptable}

\item{alpha}{the significance of the default hypothesis test, which is H0 : μt = μc vs H1 : μt != μc; The default value is 0.05}

\item{w}{the borrowing weight; we choose the optimal w that minimizes the variance as the default value, which equals \eqn{\frac{cov}{x2_var}}}

\item{W}{the capital W represents the overall test statistics, which includes the external control data if x2_mean falls within the range between -delta and delta}
}
\value{
\code{Calibration1()} returns (1) W_var: the variance of overall test statistics; (2) rejectNull: whether we should reject the null hypothesis
}
\description{
This calibration method estimate the variance of the overall test statistics based on several parameters
}
\details{
\code{Calibration1()} estimates the variance of the overall test statistics W based on several parameters and conclude a decision regarding to reject the null hypothesis or not.
 The default hypothesis test is a two-sided test with null hypothesis and alternative hypothesis equaling to H0 : μt = μc vs H1 : μt != μc
}
\seealso{
toBeAdded
}
