% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration2.R
\name{Calibration2}
\alias{Calibration2}
\title{Calibration method 2 (adjusted nominal testing level)}
\usage{
Calibration2(x1_mean, x2_mean, x1_var, x2_var, cov, delta, alpha_p = 0.05, w = cov/x2_var, W = x1_mean - w*ifelse(abs(x2_mean) < delta, x2_mean, 0))
}
\arguments{
\item{x1_mean}{the sample mean of the treatment effect estimation (mean difference between the treatment arm and control arm)}

\item{x2_mean}{the sample mean that evaluate the heterogeneity between the control arm and external control arm}

\item{x1_var}{the estimated variance of x1}

\item{x2_var}{the estimated variance of x2}

\item{cov}{the covariance between x1 and x2, which is the variance of the endpoints of the control arm}

\item{delta}{the equivalence margin, which is the maximum difference that can be treated as clinically acceptable}

\item{alpha}{the significance of the default hypothesis test, which is H0 : μt = μc vs H1 : μt != μc; The default value is 0.05}

\item{w}{the borrowing weight; we choose the optimal w that minimizes the variance as the default value, which equals cov/x2_var}

\item{W}{the capital W represents the overall test statistics, which includes the external control data if x2_mean falls within the range between -delta and delta}
}
\value{
\code{Calibration2()} returns (1) cutoffValue: the common cutoff value; (2) rejectNull: whether we should reject the null hypothesis
}
\description{
This calibration method recalculated an adjusted nominal testing level adjusted_alpha that makes the actual testing level equals alpha.
The adjusted common cutoff value would be (1-adjusted_alpha/2)-th quantile of standard normal distribution
}
\details{
\code{Calibration2()} calculate the adjusted nominal testing level that makes the actual testing level equals alpha based on several parameters and
conclude a decision regarding to reject the null hypothesis or not. The default hypothesis test is a two-sided test with null hypothesis and alternative hypothesis equaling to H0 : μt = μc vs H1 : μt != μc
}
\seealso{
toBeAdded
}
