% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration3.R
\name{Calibration3}
\alias{Calibration3}
\title{Calibration method 3 (adjusted cutoff values considering borrowing and non-borrowing cases)}
\usage{
Calibration3(x1_mean, x2_mean, x1_var, x2_var, cov, delta, p1, alpha_p = 0.05, w = cov/x2_var, W = x1_mean - w*ifelse(abs(x2_mean) < delta, x2_mean, 0))
}
\arguments{
\item{x1_mean}{the sample mean of the treatment effect estimation (mean difference between the treatment arm and control arm)}

\item{x2_mean}{the sample mean that evaluate the heterogeneity between the control arm and external control arm}

\item{x1_var}{the estimated variance of x1}

\item{x2_var}{the estimated variance of x2}

\item{cov}{the covariance between x1 and x2, which is the variance of the endpoints of the control arm}

\item{delta}{the equivalence margin, which is the maximum difference that can be treated as clinically acceptable}

\item{p1}{the proportion of type I error spitted to borrowing case; p1 can be an array to test multiple split proportion simultaneously}

\item{alpha}{the significance of the default hypothesis test, which is H0 : μt = μc vs H1 : μt != μc; The default value is 0.05}

\item{w}{the borrowing weight; we choose the optimal w that minimizes the variance as the default value, which equals cov/x2_var}

\item{W}{the capital W represents the overall test statistics, which includes the external control data if x2_mean falls within the range between -delta and delta}
}
\value{
\code{Calibration3()} returns (1) cutoffValue1: the cutoff value for the case when the equivalence between the current control arm and external control arm is established (borrowing);
 (2) cutoffValue2: the cutoff value for the case when the equivalence between the current control arm and external control arm is not established (no borrowing);
 (3) rejectNull: whether we should reject the null hypothesis of the primary test
}
\description{
This calibration method recalculated an adjusted nominal testing level and split it to borrowing and non-borrowing cases. Two separate
cutoff values will be returned in this case. We determine whether to reject the null hypothesis of the primary test based on each case and its
corresponding critical value.
}
\details{
\code{Calibration3()} calculate two separate cutoff values based on the conclusion of equivalence test and conclude a decision regarding to reject the
null hypothesis or not considering borrowing and non-borrowing cases. The default hypothesis test is a two-sided test with null hypothesis and
alternative hypothesis equaling to H0 : μt = μc vs H1 : μt != μc
}
\seealso{
toBeAdded
}
